+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/CameraAndroidCameraxPlugin.java
@@ -134,6 +134,8 @@ public final class CameraAndroidCameraxPlugin implements FlutterPlugin, Activity
         binaryMessenger, new FocusMeteringResultHostApiImpl(instanceManager));
     meteringPointHostApiImpl = new MeteringPointHostApiImpl(instanceManager);
     GeneratedCameraXLibrary.MeteringPointHostApi.setup(binaryMessenger, meteringPointHostApiImpl);
+    GeneratedCameraXLibrary.ResolutionFilterHostApi.setup(
+        binaryMessenger, new ResolutionFilterHostApiImpl(instanceManager));
   }
 
   @Override