+++ packages/pigeon/platform_tests/alternate_language_test_plugin/ios/Classes/AlternateLanguageTestPlugin.m
@@ -44,6 +44,12 @@
   return everything;
 }
 
+- (nullable FLTAllNullableTypesWithoutRecursion *)
+    echoAllNullableTypesWithoutRecursion:(nullable FLTAllNullableTypesWithoutRecursion *)everything
+                                   error:(FlutterError *_Nullable __autoreleasing *_Nonnull)error {
+  return everything;
+}
+
 - (nullable id)throwErrorWithError:(FlutterError *_Nullable *_Nonnull)error {
   *error = [FlutterError errorWithCode:@"An error" message:nil details:nil];
   return nil;
@@ -129,7 +135,9 @@
                                    error:(FlutterError *_Nullable *_Nonnull)error {
   FLTAllNullableTypes *innerObject = [[FLTAllNullableTypes alloc] init];
   innerObject.aNullableString = nullableString;
-  return [FLTAllClassesWrapper makeWithAllNullableTypes:innerObject allTypes:nil];
+  return [FLTAllClassesWrapper makeWithAllNullableTypes:innerObject
+                       allNullableTypesWithoutRecursion:nil
+                                               allTypes:nil];
 }
 
 - (nullable FLTAllNullableTypes *)
@@ -138,6 +146,21 @@
                            aString:(nullable NSString *)aNullableString
                              error:(FlutterError *_Nullable *_Nonnull)error {
   FLTAllNullableTypes *someTypes = [[FLTAllNullableTypes alloc] init];
+  someTypes.aNullableBool = aNullableBool;
+  someTypes.aNullableInt = aNullableInt;
+  someTypes.aNullableString = aNullableString;
+  return someTypes;
+}
+
+- (nullable FLTAllNullableTypesWithoutRecursion *)
+    sendMultipleNullableTypesWithoutRecursionABool:(nullable NSNumber *)aNullableBool
+                                             anInt:(nullable NSNumber *)aNullableInt
+                                           aString:(nullable NSString *)aNullableString
+                                             error:
+                                                 (FlutterError *_Nullable __autoreleasing *_Nonnull)
+                                                     error {
+  FLTAllNullableTypesWithoutRecursion *someTypes =
+      [[FLTAllNullableTypesWithoutRecursion alloc] init];
   someTypes.aNullableBool = aNullableBool;
   someTypes.aNullableInt = aNullableInt;
   someTypes.aNullableString = aNullableString;
@@ -226,6 +249,16 @@
 - (void)echoAsyncNullableAllNullableTypes:(nullable FLTAllNullableTypes *)everything
                                completion:(void (^)(FLTAllNullableTypes *_Nullable,
                                                     FlutterError *_Nullable))completion {
+  completion(everything, nil);
+}
+
+- (void)
+    echoAsyncNullableAllNullableTypesWithoutRecursion:
+        (nullable FLTAllNullableTypesWithoutRecursion *)everything
+                                           completion:
+                                               (nonnull void (^)(
+                                                   FLTAllNullableTypesWithoutRecursion *_Nullable,
+                                                   FlutterError *_Nullable))completion {
   completion(everything, nil);
 }
 
@@ -367,6 +400,25 @@
                           }];
 }
 
+- (void)callFlutterSendMultipleNullableTypesWithoutRecursionABool:(nullable NSNumber *)aNullableBool
+                                                            anInt:(nullable NSNumber *)aNullableInt
+                                                          aString:
+                                                              (nullable NSString *)aNullableString
+                                                       completion:
+                                                           (nonnull void (^)(
+                                                               FLTAllNullableTypesWithoutRecursion
+                                                                   *_Nullable,
+                                                               FlutterError *_Nullable))completion {
+  [self.flutterAPI
+      sendMultipleNullableTypesWithoutRecursionABool:aNullableBool
+                                               anInt:aNullableInt
+                                             aString:aNullableString
+                                          completion:^(FLTAllNullableTypesWithoutRecursion *value,
+                                                       FlutterError *error) {
+                                            completion(value, error);
+                                          }];
+}
+
 - (void)callFlutterEchoBool:(BOOL)aBool
                  completion:(void (^)(NSNumber *_Nullable, FlutterError *_Nullable))completion {
   [self.flutterAPI echoBool:aBool
@@ -440,6 +492,20 @@
                              completion:^(FLTAllNullableTypes *value, FlutterError *error) {
                                completion(value, error);
                              }];
+}
+
+- (void)callFlutterEchoAllNullableTypesWithoutRecursion:
+            (nullable FLTAllNullableTypesWithoutRecursion *)everything
+                                             completion:
+                                                 (nonnull void (^)(
+                                                     FLTAllNullableTypesWithoutRecursion *_Nullable,
+                                                     FlutterError *_Nullable))completion {
+  [self.flutterAPI
+      echoAllNullableTypesWithoutRecursion:everything
+                                completion:^(FLTAllNullableTypesWithoutRecursion *value,
+                                             FlutterError *error) {
+                                  completion(value, error);
+                                }];
 }
 
 - (void)callFlutterEchoNullableBool:(nullable NSNumber *)aBool