+++ packages/pigeon/platform_tests/test_plugin/macos/Classes/TestPlugin.swift
@@ -42,6 +42,11 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
   func echo(_ everything: AllNullableTypes?) -> AllNullableTypes? {
     return everything
   }
+  func echo(_ everything: AllNullableTypesWithoutRecursion?) throws
+    -> AllNullableTypesWithoutRecursion?
+  {
+    return everything
+  }
 
   func throwError() throws -> Any? {
     throw FlutterError(code: "code", message: "message", details: "details")
@@ -107,6 +112,14 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     aBool aNullableBool: Bool?, anInt aNullableInt: Int64?, aString aNullableString: String?
   ) -> AllNullableTypes {
     let someThings = AllNullableTypes(
+      aNullableBool: aNullableBool, aNullableInt: aNullableInt, aNullableString: aNullableString)
+    return someThings
+  }
+
+  func sendMultipleNullableTypesWithoutRecursion(
+    aBool aNullableBool: Bool?, anInt aNullableInt: Int64?, aString aNullableString: String?
+  ) throws -> AllNullableTypesWithoutRecursion {
+    let someThings = AllNullableTypesWithoutRecursion(
       aNullableBool: aNullableBool, aNullableInt: aNullableInt, aNullableString: aNullableString)
     return someThings
   }
@@ -190,6 +203,13 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
   func echoAsync(
     _ everything: AllNullableTypes?,
     completion: @escaping (Result<AllNullableTypes?, Error>) -> Void
+  ) {
+    completion(.success(everything))
+  }
+
+  func echoAsync(
+    _ everything: AllNullableTypesWithoutRecursion?,
+    completion: @escaping (Result<AllNullableTypesWithoutRecursion?, Error>) -> Void
   ) {
     completion(.success(everything))
   }
@@ -339,6 +359,20 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     }
   }
 
+  func callFlutterEcho(
+    _ everything: AllNullableTypesWithoutRecursion?,
+    completion: @escaping (Result<AllNullableTypesWithoutRecursion?, Error>) -> Void
+  ) {
+    flutterAPI.echoNullable(everything) { response in
+      switch response {
+      case .success(let res):
+        completion(.success(res))
+      case .failure(let error):
+        completion(.failure(error))
+      }
+    }
+  }
+
   func callFlutterSendMultipleNullableTypes(
     aBool aNullableBool: Bool?,
     anInt aNullableInt: Int64?,
@@ -346,6 +380,24 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
     completion: @escaping (Result<AllNullableTypes, Error>) -> Void
   ) {
     flutterAPI.sendMultipleNullableTypes(
+      aBool: aNullableBool,
+      anInt: aNullableInt,
+      aString: aNullableString
+    ) { response in
+      switch response {
+      case .success(let res):
+        completion(.success(res))
+      case .failure(let error):
+        completion(.failure(error))
+      }
+    }
+  }
+
+  func callFlutterSendMultipleNullableTypesWithoutRecursion(
+    aBool aNullableBool: Bool?, anInt aNullableInt: Int64?, aString aNullableString: String?,
+    completion: @escaping (Result<AllNullableTypesWithoutRecursion, Error>) -> Void
+  ) {
+    flutterAPI.sendMultipleNullableTypesWithoutRecursion(
       aBool: aNullableBool,
       anInt: aNullableInt,
       aString: aNullableString